tests/submission/PhilipTaylor/canvas/toDataURL.png.primarycolours.html

<!DOCTYPE html>
<html><head><meta charset="utf-8">
<title>Canvas test: toDataURL.png.primarycolours</title>
<meta content="Philip Taylor" name="author">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="../common/canvas-tests.js"></script>
<link href="../common/canvas-tests.css" rel="stylesheet">
</head><body class="show_output">

<h1><a href="index.toDataURL.html">toDataURL</a>.<a href="index.toDataURL.png.html">png</a>.primarycolours</h1>
<p class="desc">toDataURL with PNG handles simple colours correctly</p>

<div class="refs">Spec references:
<ul>
<li><a href="../annotated-spec/canvas.html#testrefs.toDataURL.png">toDataURL.png</a>
</li></ul>
</div>

<p class="output">Actual output:</p>
<canvas class="output" height="50" id="c" width="100"><p class="fallback">FAIL (fallback content)</p></canvas>
<p class="output expectedtext">Expected output:</p><p><img alt="" class="output expected" id="expected" src="toDataURL.png.primarycolours.png">
</p><ul id="d"></ul>
<script>
var t = async_test("toDataURL with PNG handles simple colours correctly");
_addTest(function(canvas, ctx) {

ctx.fillStyle = '#ff0';
ctx.fillRect(0, 0, 25, 40);
ctx.fillStyle = '#0ff';
ctx.fillRect(25, 0, 50, 40);
ctx.fillStyle = '#00f';
ctx.fillRect(75, 0, 25, 40);
ctx.fillStyle = '#fff';
ctx.fillRect(0, 40, 100, 10);
var data = canvas.toDataURL();
ctx.fillStyle = '#f00';
ctx.fillRect(0, 0, 100, 50);
var img = new Image();
deferTest();
img.onload = wrapFunction(function ()
{
    ctx.drawImage(img, 0, 0);
    
    _assertPixel(canvas, 12,20, 255,255,0,255, "12,20", "255,255,0,255");
    _assertPixel(canvas, 50,20, 0,255,255,255, "50,20", "0,255,255,255");
    _assertPixel(canvas, 87,20, 0,0,255,255, "87,20", "0,0,255,255");
    _assertPixel(canvas, 50,45, 255,255,255,255, "50,45", "255,255,255,255");
});
img.src = data;


});
</script>

<div id="log"></div>
</body></html>
